Remove some ee imports
