+++ backend/ee/onyx/background/celery/tasks/doc_permission_syncing/tasks.py
