+++ backend/ee/onyx/background/celery/tasks/external_group_syncing/group_sync_utils.py
